# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "FDDT — Fraud Document Detection Tool API",
    "description": "FDDT (Fraud Document Detection Tool) backend. Environment: local.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/platform/companies": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "List companies",
        "operationId": "list_companies_platform_companies_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/CompanyResponse"
                  },
                  "type": "array",
                  "title": "Response List Companies Platform Companies Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Create a company",
        "description": "Creates a client company and copies every active platform rule template into its own risk rules, plus default tier thresholds (its own independent copy, which its Reviewer L2s can then tune; later template edits don't affect it). The issuer registry starts empty. Users are added separately (Settings > Users).",
        "operationId": "create_company_platform_companies_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          },
          "409": {
            "description": "A company with this name already exists."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/platform/companies/{company_id}": {
      "patch": {
        "tags": [
          "platform"
        ],
        "summary": "Rename, suspend/reactivate, or change a company's upload limits",
        "description": "`is_active=false` suspends the company: its users can no longer sign in, and existing tokens stop working on the next request. Data is kept. `max_file_size_mb` / `max_zip_size_mb` change the company's upload limits (any positive whole number of MB; no upper ceiling), effective on its next upload with no new sign-in. Every change is written to the platform audit log as `company_updated` with old → new values.",
        "operationId": "update_company_platform_companies__company_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompanyUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CompanyResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          },
          "404": {
            "description": "No such company."
          },
          "409": {
            "description": "Name taken."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/platform/risk-rule-templates": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "List the platform risk-rule template",
        "description": "The default rule set copied into every NEW company's own risk rules at creation time. Editing it never changes companies that already exist.",
        "operationId": "list_rule_templates_platform_risk_rule_templates_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/RuleTemplateResponse"
                  },
                  "type": "array",
                  "title": "Response List Rule Templates Platform Risk Rule Templates Get"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Add a rule to the platform template",
        "description": "Same rule builder as a company's Add-rule form. Applies to companies created from now on.",
        "operationId": "create_rule_template_platform_risk_rule_templates_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RiskRuleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleTemplateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          },
          "409": {
            "description": "A template rule with this rule_id already exists."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/platform/risk-rule-templates/{rule_id}": {
      "patch": {
        "tags": [
          "platform"
        ],
        "summary": "Edit a platform template rule",
        "description": "Changes weight, severity, wording or active flag. Existing companies keep their own copies unchanged; only companies created afterwards start from the new values.",
        "operationId": "update_rule_template_platform_risk_rule_templates__rule_id__patch",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "rule_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Rule Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RuleTemplateUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RuleTemplateResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          },
          "404": {
            "description": "No such template rule."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/platform/usage": {
      "get": {
        "tags": [
          "platform"
        ],
        "summary": "Billing & usage by company",
        "description": "One row per company for the period: cases created, documents uploaded, files stored (documents + generated reports) and bytes stored during the period, plus each company's current all-time storage. Read from incrementally-maintained daily counters (company_usage_stats), not live scans of the case tables; a nightly job reconciles them against the source data. Periods use UTC days.",
        "operationId": "get_usage_platform_usage_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "period",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "this_month",
                "last_month",
                "last_30_days",
                "this_year",
                "all_time",
                "custom"
              ],
              "type": "string",
              "default": "this_month",
              "title": "Period"
            }
          },
          {
            "name": "start",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "custom period only",
              "title": "Start"
            },
            "description": "custom period only"
          },
          {
            "name": "end",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date"
                },
                {
                  "type": "null"
                }
              ],
              "description": "custom period only",
              "title": "End"
            },
            "description": "custom period only"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/UsageResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/platform/usage/reconcile": {
      "post": {
        "tags": [
          "platform"
        ],
        "summary": "Reconcile usage counters now",
        "description": "Recomputes every company's daily counters from the cases/documents/reports tables and corrects any drift (the same job runs nightly). Returns what was corrected; a correction is also written to the platform audit log.",
        "operationId": "reconcile_now_platform_usage_reconcile_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReconcileResponse"
                }
              }
            }
          },
          "401": {
            "description": "Missing, invalid or expired bearer token, or the user is inactive."
          },
          "403": {
            "description": "Requires the platform admin role."
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CompanyResponse": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "is_active": {
            "type": "boolean",
            "title": "Is Active"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "user_count": {
            "type": "integer",
            "title": "User Count"
          },
          "max_file_size_mb": {
            "type": "integer",
            "title": "Max File Size Mb",
            "description": "Largest document this company may upload, in MB."
          },
          "max_zip_size_mb": {
            "type": "integer",
            "title": "Max Zip Size Mb",
            "description": "Largest bulk-upload zip this company may upload, in MB."
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_active",
          "created_at",
          "user_count",
          "max_file_size_mb",
          "max_zip_size_mb"
        ],
        "title": "CompanyResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}